pyproject.toml
[tool.poetry]
name = "weather-forecast"
version = "0.1.0"
description = ""
authors = ["Пользователь <yaroslav-rusin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-bootstrap-v5 = "^1.0.11"
django-select2 = "^8.2.0"
requests = "^2.32.3"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
numpy = "^2.0.0"
pandas = "^2.2.2"
djangorestframework = "^3.15.2"
django = "4.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
